requirements.txt
asgiref==3.8.1
Django==4.2.12
django-allauth==0.62.1
djangorestframework==3.15.1
sqlparse==0.5.0
typing-extensions==4.11.0
